Function

Writes an executes a Makefile to distribute remake targets over simultaneous processes. Use the help_parallelRemake function to get more help.

makefile(targets = "all", remakefiles = "remake.yml", prepend = NULL,
  remake_args = list(verbose = TRUE), run = TRUE, command = "make",
  args = character(0))

Arguments

targets

character vector of targets to make

remakefiles

Character vector of paths to input remake files.

prepend

Character vector of lines to prepend to the Makefile.

remake_args

Named list of additional arguments to remake::make.

run

logical, whether to actually run the Makefile or just write it.

command

character scalar, command to run to execute the Makefile. command and args will be used to call system2(command = command, args = args) to run the Makefile. For example, to execute the Makefile using 4 parallel jobs while suppressing output to the console, use makefile(..., command = "make", args = c("--jobs=4", "-s")).

args

character vector of arguments to the specified command. command and args will be used to call system2(command = command, args = args) to run the Makefile. For example, to execute the Makefile using 4 parallel jobs while suppressing output to the console, use makefile(..., command = "make", args = c("--jobs=4", "-s")).

Details

Use the help_parallelRemake function to get more help.

See also

help_parallelRemake