Function

Writes a master Makefile that can distribute remake workflows over simultaneous processes. Use the help_parallelRemake function to get more help.

write_makefile(makefile = "Makefile", remakefiles = "remake.yml",
  begin = NULL, clean = NULL, remake_args = list())

Arguments

makefile

Name of the Makefile to write.

remakefiles

Character vector of paths to input remake files.

begin

Character vector of lines to prepend to the Makefile.

clean

Character vector of commands to add to the clean rule.

remake_args

Named list of additional arguments to remake::make. You cannot set target_names or remake_file this way because those names are already reserved.

Details

Use the help_parallelRemake function to get more help.

See also

help_parallelRemake